Some stuff
Very buggy but well

+++ src/sero583/pacman/Bot.java
@@ -3,7 +3,7 @@ package sero583.pacman;
 import java.awt.Color;
 
 public class Bot extends Entity {
-	public Bot(int id, Color color, int x, int y) {
+	public Bot(int id, Color color, double x, double y) {
 		super(id, color, x, y);
 	}
